Updates for mobile masthead and breadcrumbs.
Added CSS to allow a department level link in the masthead beneath the existing
brand link.  Wraps just before the collapsible menu toggle if long enough to
need it.  Also, used an unsharp mask to make the brand link look crisper and
brighter.  This change requires duplication of the department link to appear
in both wide (sm - lg) and narrow (xs) viewports.  Breadcrumbs are back to using
Glyphicon Halflings chevron due to poor support for the unicode chevron.